[feat] MarvelCharacter모델 목업데이터 #6
